LC-count-pairs-that-form-a-complete-day-ii [Runtime: 3 ms; Memory: 97.3 MB]

+++ array/count-pairs-that-form-a-complete-day-ii.java
@@ -1,6 +1,6 @@
 class Solution {
     public long countCompleteDayPairs(int[] hours) {
-        int ans = 0;
+        long ans = 0;
         if (hours.length == 0 ){
             return 0;
         }
